Separation of jvm-specific code from DDLTest
